refactor: common module integrates onion layers
The 3 main layers from onion arquitecture (DOMAIN, APPLICATION, INFRASTRUCTURE) was implemented, to follow project arquitecture & maintability.
