+++ src/app.module.ts
@@ -6,8 +6,8 @@ import { PrismaModule } from './prisma/prisma.module';
 import { UserModule } from './users/user.module';
 import { CategoryModule } from './categories/category.module';
 import { CustomerModule } from './customers/customer.module';
-import { CorrelationIdMiddleware } from './common/middlewares/correlation-id/correlation-id.middleware';
-import { loggerFactory } from './common/config/logger.factory';
+import { CorrelationIdMiddleware } from './common/application/middlewares/correlation-id/correlation-id.middleware';
+import { loggerFactory } from './common/application/config/logger.factory';
 import { AuthModule } from './auth/auth.module';
 import { InventoryModule } from './inventory/inventory.module';
 import { ProductModule } from './products/product.module';
