+++ src/categories/application/repositories/category.repository.spec.ts
@@ -1,5 +1,4 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { Category, Prisma } from '@prisma/client';
 
 import { CategoryRepository } from './category.repository';
 import { PrismaProvider } from '../../../prisma/providers/prisma.provider';
