+++ src/categories/infrastructure/controllers/category.controller.ts
@@ -3,7 +3,7 @@ import { ApiBearerAuth, ApiOperation, ApiParam, ApiResponse, ApiTags } from '@ne
 import { Category } from '@prisma/client';
 
 import { CategoryService } from '../../application/services/category.service';
-import { uuidPipeOptions } from 'src/common/config/uuid-pipe.config';
+import { uuidPipeOptions } from '../../../common/application/config';
 import { CategoryEntity } from '../../domain/entities/category.entity';
 
 @ApiBearerAuth()
