+++ src/common/application/decorators/index.ts
@@ -1,0 +1,1 @@
+export * from './public.decorator';
