+++ src/common/application/decorators/public.decorator.ts
@@ -1,5 +1,5 @@
 import { CustomDecorator, SetMetadata } from "@nestjs/common";
 
-import { PUBLIC_KEY } from "../constants/public-key.decorator";
+import { PUBLIC_KEY } from "../../domain/constants";
 
 export const PublicAccess = (): CustomDecorator<string> => SetMetadata(PUBLIC_KEY, true);
