--- src/common/config/index.ts
@@ -1,3 +1,0 @@
-export { appConfig } from './app.config';
-export { validationOptions } from './validation-pipe.config';
-export { swaggerOptions } from './swagger.config';
