+++ src/common/domain/constants/index.ts
@@ -1,0 +1,1 @@
+export * from './public-decorator.key';
