+++ src/common/domain/types/error.type.ts
