+++ src/common/domain/types/payload.type.ts
