+++ src/customers/application/repositories/customer.repository.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@nestjs/common';
 import { PrismaProvider } from '../../../prisma/providers/prisma.provider';
 import { CustomerEntity } from '../../domain/entities/customer.entity';
 import { SelectedFields } from '../../domain/types';
-import { Pagination } from '../../../common/types/pagination.type';
+import { Pagination } from '../../../common/domain/types';
 
 @Injectable()
 export class CustomerRepository {
