+++ src/products/application/services/product.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 
 import { ProductRepository } from '../repositories/product.repository';
 import { SelectedFields } from '../../domain/types';
-import { Pagination } from "../../../common/types";
+import { Pagination } from "../../../common/domain/types/pagination.type";
 import { SearchTerms } from '../../domain/types/search-terms.type';
 import { ProductEntity } from '../../domain/entities/product.entity';
