+++ src/products/domain/types/product-queries.type.ts
@@ -1,4 +1,3 @@
-import { Pagination } from "src/common/types"
 import { SearchTerms } from "./search-terms.type"
 
 export type ProductQueries = {
